LayoutTests/imported/w3c/web-platform-tests/svg/animations/animateMotion-line-expected.txt


FAIL Test for checking position of the svg element when animateMotion's mpath is a line assert_approx_equals: expected 50 +/- 1 but got 0
